Complete module privacy refactor
