+++ src/response.rs
@@ -1,9 +1,13 @@
 //! Stream the updates appropriate for a given `User`/`timeline` pair from Redis.
 
-pub use crate::event::Event;
+pub use event::Event;
 pub use redis::Manager as RedisManager;
 pub use stream::{Sse as SseStream, Ws as WsStream};
 
+pub(self) use event::err::Event as EventErr;
+pub(self) use event::Payload;
+
+pub(crate) mod event;
 mod redis;
 mod stream;
