+++ src/response/event/checked_event/account.rs
@@ -1,4 +1,5 @@
-use super::{emoji::Emoji, id::Id, visibility::Visibility};
+use super::{emoji::Emoji, visibility::Visibility};
+use crate::Id;
 use serde::{Deserialize, Serialize};
 
 #[serde(deny_unknown_fields)]
