+++ src/response/event/checked_event/announcement.rs
