+++ src/response/event/checked_event/announcement_reaction.rs
