+++ src/response/event/checked_event/emoji.rs
