+++ src/response/stream.rs
@@ -1,5 +1,7 @@
 pub use sse::Sse;
 pub use ws::Ws;
 
+pub(self) use super::{Event, Payload};
+
 mod sse;
 mod ws;
